refactor - Refactored the some components location & the dropdown component code
